+++ gui/PupitreWindow.cpp
@@ -17,7 +17,7 @@ PupitreWindow::PupitreWindow(std::vector<Pupitre> pupitres) : QWidget() {
 
 	QVBoxLayout *vBox = new QVBoxLayout;
 
-	for (int i = 0; i < pupitres.size(); i++) {
+	for (size_t i = 0; i < pupitres.size(); i++) {
 		std::string nameInstru = pupitres[i].getInstrument().getname();
 		QRadioButton *b = new QRadioButton(QString(nameInstru.c_str()));
 		buttons.push_back(b);
@@ -36,7 +36,7 @@ PupitreWindow::PupitreWindow(std::vector<Pupitre> pupitres) : QWidget() {
 }
 
 void PupitreWindow::choosePupitre() {
-	for (int i = 0; i < buttons.size(); i++) {
+	for (size_t i = 0; i < buttons.size(); i++) {
 		if (buttons[i]->isChecked()) {
 			Pupitre p = this->pupitres[i];
 			std::cout << "Instrument avec lequel vous allez jouer : "<< p.getInstrument().getname() << std::endl;
